requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==41.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
discord==2.2.3
discord.py==2.2.3
dj-rest-auth==5.0.1
Django==4.2.4
django-allauth==0.57.0
django-autoslug==1.9.9
djangorestframework==3.14.0
docopt==0.6.2
efficientnet-pytorch==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
flatbuffers==22.12.6
fonttools==4.28.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==21.2.0
h5py==3.7.0
idna==2.10
image==1.5.33
imageio==2.23.0
ipykernel==6.6.0
ipython==7.30.1
ipywidgets==8.0.4
jedi==0.18.1
joblib==1.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-widgets==3.0.5
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy_loader==0.2
libclang==14.0.6
librosa==0.10.0.post2
llvmlite==0.39.1
lxml==4.6.5
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.0.28
nest-asyncio==1.5.4
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==8.4.0
pooch==1.6.0
prompt-toolkit==3.0.24
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.11.0
Pygments==2.10.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==302
pywin32-ctypes==0.2.2
pyzmq==22.3.0
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0.post1
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.1
Unidecode==1.3.6
urllib3==1.25.11
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.2.2
whitenoise==6.5.0
widgetsnbextension==4.0.5
wrapt==1.14.1
xlrd==2.0.1
yarl==1.9.2
